requirements.txt
google-generativeai
python-dotenv
pandas
numpy
matplotlib
seaborn
